таблицы

+++ src/main/java/ru/plot/entity/Okv.java
@@ -3,10 +3,8 @@ package ru.plot.entity;
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
+import javax.persistence.*;
+import java.math.BigInteger;
 
 @Getter
 @Setter
@@ -16,7 +14,7 @@ public class Okv {
 
     @Id
     @Column(name = "id_okv")
-    private Long id;
+    private BigInteger id;
 
     @Column(name = "charcode")
     private String charcode;
